# docs-site/openapi/public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feedling API",
    "version": "v1",
    "description": "HTTP API for Feedling accounts, encrypted chat, agent memory, model routing, perception, and proactive experiences."
  },
  "servers": [
    {
      "url": "https://api.feedling.app",
      "description": "Production"
    },
    {
      "url": "https://test-api.feedling.app",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "RuntimeTokenAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/access/claim-token": {
      "post": {
        "summary": "Access Link Token Claim",
        "operationId": "access_link_token_claim_v1_access_claim_token_post",
        "responses": {
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Link token consumed and an additional API key issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedApiKeyResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimTokenRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented",
        "description": "Consume a one-time link token and issue an additional API key. Existing keys remain active."
      }
    },
    "/v1/access/link-token": {
      "post": {
        "summary": "Access Link Token Create",
        "operationId": "access_link_token_create_v1_access_link_token_post",
        "responses": {
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "One-time link token created. It expires after the returned TTL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkTokenResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkTokenRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented"
      }
    },
    "/v1/access/modes": {
      "get": {
        "summary": "Access Modes",
        "operationId": "access_modes_v1_access_modes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "x-feedling-contract-level": "documented"
      }
    },
    "/v1/access/modes/switch": {
      "post": {
        "summary": "Access Modes Switch",
        "operationId": "access_modes_switch_v1_access_modes_switch_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreeFormJsonObject"
              }
            }
          }
        },
        "x-feedling-contract-level": "compatibility"
      }
    },
    "/v1/account/recover/challenge": {
      "post": {
        "summary": "Account Recover Challenge",
        "operationId": "account_recover_challenge_v1_account_recover_challenge_post",
        "responses": {
          "200": {
            "description": "Single-use encrypted recovery challenge created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoverChallengeResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverChallengeRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented"
      }
    },
    "/v1/account/recover/verify": {
      "post": {
        "summary": "Account Recover Verify",
        "operationId": "account_recover_verify_v1_account_recover_verify_post",
        "responses": {
          "200": {
            "description": "Challenge verified and an additional API key issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedApiKeyResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverVerifyRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented",
        "description": "Verify keypair possession and issue an additional API key for the existing account. Existing keys remain active."
      }
    },
...
    "/v1/onboarding/route": {
      "get": {
        "summary": "Onboarding Route Get",
        "operationId": "onboarding_route_get_v1_onboarding_route_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Onboarding"
        ],
        "x-feedling-contract-level": "documented"
      },
      "post": {
        "summary": "Onboarding Route Post",
        "operationId": "onboarding_route_post_v1_onboarding_route_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Onboarding"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreeFormJsonObject"
              }
            }
          }
        },
        "x-feedling-contract-level": "compatibility"
      }
    },
...
    "/v1/users/preferences": {
      "post": {
        "summary": "Users Set Preferences",
        "operationId": "users_set_preferences_v1_users_preferences_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreeFormJsonObject"
              }
            }
          }
        },
        "x-feedling-contract-level": "compatibility"
      }
    },
...
    "/v1/users/register": {
      "post": {
        "summary": "Users Register",
        "operationId": "users_register_v1_users_register_post",
        "responses": {
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "User created; the plaintext API key is returned once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "security": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "x-feedling-contract-level": "documented",
        "description": "Create a Feedling user and issue its first API key. The key is returned once and this operation is not idempotent."
      }
    },
    "/v1/users/whoami": {
      "get": {
        "summary": "Whoami",
        "operationId": "whoami_v1_users_whoami_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericJsonResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ],
        "x-feedling-contract-level": "documented"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FreeFormJsonObject": {
        "type": "object",
        "description": "Legacy compatibility payload. Consult the operation description before sending fields.",
        "additionalProperties": true
      },
      "GenericJsonResponse": {
        "type": "object",
        "description": "JSON response object. Endpoint-specific fields may be added compatibly.",
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "description": "Machine-readable error slug, or the legacy structured MCP error.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "required": [
                  "kind"
                ],
                "properties": {
                  "kind": {
                    "type": "string"
                  },
                  "detail": {
                    "type": "string"
                  }
                },
                "additionalProperties": true
              }
            ]
          },
          "detail": {
            "description": "Human-readable context or structured validation details; do not branch on prose.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {}
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "blame": {
            "type": "string",
            "enum": [
              "user_provider",
              "provider_transient",
              "system"
            ],
            "description": "Stable responsibility classification when the endpoint can identify it."
          },
          "request_id": {
            "type": "string",
            "description": "Support correlation identifier when available."
          }
        },
        "additionalProperties": true,
        "example": {
          "error": "invalid_payload",
          "detail": "The request body is invalid",
          "request_id": "req_abc123"
        }
      },
      "EncryptedEnvelope": {
        "type": "object",
        "required": [
          "body_ct",
          "nonce",
          "K_user",
          "visibility",
          "owner_user_id"
        ],
        "properties": {
          "v": {
            "type": "integer",
            "const": 1,
            "default": 1
          },
          "id": {
            "type": "string"
          },
          "body_ct": {
            "type": "string",
            "description": "Base64 ciphertext."
          },
          "nonce": {
            "type": "string",
            "description": "Base64 nonce."
          },
          "K_user": {
            "type": "string",
            "description": "Content key wrapped for the user."
          },
          "K_enclave": {
            "type": "string",
            "description": "Content key wrapped for the enclave; required for shared visibility."
          },
          "visibility": {
            "type": "string",
            "enum": [
              "shared",
              "local_only"
            ]
          },
          "owner_user_id": {
            "type": "string"
          },
          "enclave_pk_fpr": {
            "type": "string"
          }
        },
        "if": {
          "properties": {
            "visibility": {
              "const": "shared"
            }
          },
          "required": [
            "visibility"
          ]
        },
        "then": {
          "required": [
            "K_enclave"
          ]
        },
        "additionalProperties": true,
        "example": {
          "v": 1,
          "id": "env_abc123",
          "body_ct": "BASE64_CIPHERTEXT",
          "nonce": "BASE64_NONCE",
          "K_user": "BASE64_KEY_WRAPPED_TO_USER",
          "K_enclave": "BASE64_KEY_WRAPPED_TO_ENCLAVE",
          "visibility": "shared",
          "owner_user_id": "usr_0123456789abcdef"
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "public_key": {
            "type": "string",
            "description": "Base64 content-encryption public key."
          },
          "archive_language": {
            "type": "string",
            "example": "en"
          },
          "access_mode": {
            "type": "string",
            "enum": [
              "resident",
              "model_api",
              "official_import"
            ],
            "default": "official_import"
          },
          "label": {
            "type": "string",
            "maxLength": 80
          }
        },
        "additionalProperties": false
      },
      "RegisterResponse": {
        "type": "object",
        "required": [
          "user_id",
          "principal_id",
          "api_key"
        ],
        "properties": {
          "user_id": {
            "type": "string"
          },
          "principal_id": {
            "type": "string"
          },
          "api_key": {
            "type": "string",
            "description": "Shown once; store it securely."
          },
          "public_key": {
            "type": "string"
          },
          "access_mode": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "LinkTokenRequest": {
        "type": "object",
        "properties": {
          "access_mode": {
            "type": "string",
            "enum": [
              "resident",
              "model_api",
              "official_import"
            ]
          },
          "label": {
            "type": "string",
            "maxLength": 80
          }
        },
        "additionalProperties": false
      },
      "ClaimTokenRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1
          },
          "label": {
            "type": "string",
            "maxLength": 80
          },
          "client_label": {
            "type": "string",
            "maxLength": 80,
            "deprecated": true
          },
          "public_key": {
            "type": "string"
          },
          "archive_language": {
            "type": "string"
          },
          "make_active": {
            "type": "boolean",
            "default": true
          }
        },
        "additionalProperties": false
      },
      "LinkTokenResponse": {
        "type": "object",
        "required": [
          "token",
          "token_id",
          "access_mode",
          "expires_at",
          "expires_in_seconds"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "One-time bearer token; handle as a secret."
          },
          "token_id": {
            "type": "string"
          },
          "access_mode": {
            "type": "string"
          },
          "route": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          },
          "expires_in_seconds": {
            "type": "integer",
            "example": 900
          },
          "claim_endpoint": {
            "type": "string",
            "const": "/v1/access/claim-token"
          }
        },
        "additionalProperties": true
      },
      "IssuedApiKeyResponse": {
        "type": "object",
        "required": [
          "user_id",
          "principal_id",
          "api_key"
        ],
        "properties": {
          "user_id": {
            "type": "string"
          },
          "principal_id": {
            "type": "string"
          },
          "api_key": {
            "type": "string",
            "description": "New key shown once; existing keys remain active."
          },
          "key_id": {
            "type": "string"
          },
          "access_mode": {
            "type": "string"
          },
          "route": {
            "type": "string"
          },
          "active_route": {
            "type": "string"
          },
          "public_key": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
      "RecoverChallengeRequest": {
        "type": "object",
        "required": [
          "public_key"
        ],
        "properties": {
          "public_key": {
            "type": "string",
            "description": "Base64 content-encryption public key."
          }
        },
        "additionalProperties": false
      },
      "RecoverVerifyRequest": {
        "type": "object",
        "required": [
          "challenge_id",
          "answer"
        ],
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "answer": {
            "type": "string",
            "description": "Plaintext recovered by decrypting the challenge envelope locally."
          }
        },
        "additionalProperties": false
      },
      "RecoverChallengeResponse": {
        "type": "object",
        "required": [
          "challenge_id",
          "envelope"
        ],
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "envelope": {
            "$ref": "#/components/schemas/EncryptedEnvelope"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived Feedling user API key."
      },
      "RuntimeTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Feedling-Runtime-Token",
        "description": "Short-lived scoped token used by hosted runtimes."
      }
    }
  }
}